requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   qrcode
cryptography==46.0.5
    # via google-auth
fastapi==0.135.1
    # via wedding-photo-contest
google-api-core==2.30.0
    # via
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth==2.49.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.25.0
    # via wedding-photo-contest
google-cloud-storage==3.9.0
    # via wedding-photo-contest
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.3.2 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
grpcio==1.78.0
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   grpcio-status
grpcio-status==1.78.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via wedding-photo-contest
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via wedding-photo-contest
jinja2==3.1.6
    # via wedding-photo-contest
markupsafe==3.0.3
    # via jinja2
pillow==12.1.1
    # via wedding-photo-contest
proto-plus==1.27.1
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==6.33.5
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.5
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via wedding-photo-contest
python-dotenv==1.2.2
    # via pydantic-settings
python-multipart==0.0.22
    # via wedding-photo-contest
qrcode==8.2
    # via wedding-photo-contest
requests==2.32.5
    # via
    #   google-api-core
    #   google-cloud-storage
sqlalchemy==2.0.48
    # via wedding-photo-contest
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.41.0
    # via wedding-photo-contest
